Rework to simulate plucking
Now, the MIDI note on event is set when the simulated string
is released, to simulate a plucked string. Each string has a
a duration, which sets how long that string will "vibrate"
(that is, how long before a MIDI note off message is sent).
Touching a vibrating string will also cause a MIDI note
off message to be sent.

Currently, all strings have the same duration. A more
accurate simulation would make the lower notes have
a longer duration.